chore: moved letalo endpoints to routes package
